requirements.txt
chromadb==1.3.5
google-generativeai==0.8.5
pandas==2.3.3
pillow==12.0.0
PyYAML==6.0.3
streamlit==1.51.0
watchdog==6.0.0

# LangChain组件（RAG层升级新增）
langchain-text-splitters==0.3.0
rank-bm25==0.2.2
sentence-transformers==3.3.1
numpy>=1.24.0